requirements.txt
anthropic==0.45.2
cryptography==44.0.0
Flask==3.1.0
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
